Finish part 2 of Get Started with Docker
Docs: https://docs.docker.com/get-started/part2/
Docker is used to create images, all-in-one executable package
containing code and environments, and run containers which are instances
of those images. So far, this project has a Dockerfile which defines the
environment of the container. I'm requiring a Python version, a project
directory in the container, the code files to be copied over to that
directory, pip dependencies from a requirements file, exposing ports,
environment variables, and a final command for Python to run my app.

Docker can build the image with `docker build --tag=friendlyhello .`
which puts the image in a local registry. It can be run with `docker run
-p 4000:80 friendlyhello` which maps the port so the app can be seen on
the host at `localhost:4000`. The run command accepts the `-d` flag to
detach the process and run in the background, and `docker container stop
<id>` will stop it. Find the id with `docker container ls`